rename folders on store
