+++ src/client/src/store/ActionType.ts
@@ -1,8 +1,8 @@
 import AzureActionType from "./azureProfileData/azureActionType";
 import ModalActionType from "./modals/modalActionType";
-import WizardContentActionType from "./templates/wizardContentActionType";
+import WizardContentActionType from "./templates/templatesActionType";
 import WizardSelectionActionType from "./selection/selectionActionType";
-import WizardInfoActionType from "./templates/wizardInfoActionType";
+import WizardInfoActionType from "./templates/templatesType";
 
 type RootAction =
   | AzureActionType
