+++ src/client/src/store/config/configType.ts
@@ -5,7 +5,7 @@ import {
 } from "../generationStatus/model";
 
 import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "./pages/model";
-import { IResetWizardAction } from "./wizardContent/model";
+import { IResetWizardAction } from "./config/model";
 
 type WizardInfoType =
   | IVersionData
