+++ src/client/src/store/selection/pages/visitedAndSetPageReducer.ts
@@ -1,6 +1,6 @@
 import { ROUTES } from "../../../utils/constants";
 import { WIZARD_INFO_TYPEKEYS } from "../../typeKeys";
-import WizardInfoType from "../../templates/wizardInfoActionType";
+import WizardInfoType from "../../templates/templatesType";
 import { IRoutes } from "./model";
 
 const initialState = {
