+++ src/client/src/store/templates/combineReducers.ts
@@ -3,11 +3,11 @@ import backendOptions from "./frameworks/backendFrameworkReducer";
 import frontendOptions from "./frameworks/frontendFrameworkReducer";
 import pageOptions from "./pages/pagesOptionsReducer";
 
-const wizardContentReducer = combineReducers({
+const templatesReducer = combineReducers({
   backendOptions,
   frontendOptions,
   pageOptions
 });
 
-export default wizardContentReducer;
-export type TemplateType = ReturnType<typeof wizardContentReducer>;
+export default templatesReducer;
+export type TemplateType = ReturnType<typeof templatesReducer>;
