+++ src/client/src/store/templates/pages/pagesOptionsReducer.ts
@@ -1,4 +1,4 @@
-import WizardContentActionType from "../wizardContentActionType";
+import WizardContentActionType from "../templatesActionType";
 import { IOption } from "../../../types/option";
 import { IResetPagesAction } from "../../selection/pages/model";
 import { WIZARD_CONTENT_TYPEKEYS } from "../../typeKeys";
