+++ src/client/src/store/templates/preview/previewReducer.ts
@@ -1,4 +1,4 @@
-import WizardContentActionType from "../wizardContentActionType";
+import WizardContentActionType from "../templatesActionType";
 import { WIZARD_CONTENT_TYPEKEYS } from "../../typeKeys";
 
 const previewStatus = (
